+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+2.5.0 (2014-05-06)
+------------------
+
+  * raised minimum PHP version to 5.4.0
+  * the following methods now except anything that can be casted to an instance of `net\stubbles\lang\types\Date` via `net\stubbles\lang\types\Date::castFrom()`
+     * `net\stubbles\input\ValueReader::asDate()`
+     * `net\stubbles\input\filter\range\DateRange::__construct()`
+     * `net\stubbles\input\filter\range\DatespanRange::__construct()`
+
+
 2.4.1 (2014-02-18)
 ------------------